requirements.txt
jieba
whoosh
openai
tiktoken
chromadb
streamlit
pandas
numpy
PyMuPDF
python-dotenv
pillow